+++ packages/app/lib/i18n/gen/strings.g.dart
@@ -4,9 +4,9 @@
 /// To regenerate, run: `dart run slang`
 ///
 /// Locales: 3
-/// Strings: 2029 (676 per locale)
+/// Strings: 2035 (678 per locale)
 ///
-/// Built on 2026-06-03 at 01:53 UTC
+/// Built on 2026-06-03 at 22:16 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import